requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
brotli-python=1.0.9=py312hcec6c5f_8
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.9.24=hecd8cb5_0
certifi=2024.8.30=py312hecd8cb5_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
diffusers=0.30.3=py312hecd8cb5_0
diffusers-base=0.30.3=py312hecd8cb5_0
diffusers-torch=0.30.3=py312hecd8cb5_0
expat=2.6.3=h6d0c2b6_0
filelock=3.13.1=py312hecd8cb5_0
freetype=2.12.1=hd8bbffd_0
fsspec=2024.6.1=py312hecd8cb5_0
huggingface_accelerate=1.0.1=py312hecd8cb5_0
huggingface_hub=0.24.6=py312hecd8cb5_0
idna=3.7=py312hecd8cb5_0
importlib-metadata=7.0.1=py312hecd8cb5_0
intel-openmp=2023.1.0=ha357a0b_43548
jinja2=3.1.4=py312hecd8cb5_1
jpeg=9e=h46256e1_3
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libffi=3.4.4=hecd8cb5_1
libjpeg-turbo=2.0.0=hca72f7f_0
libpng=1.6.39=h6c40b1e_0
libtiff=4.5.1=hcec6c5f_0
libwebp-base=1.3.2=h46256e1_1
lz4-c=1.9.4=hcec6c5f_1
markupsafe=2.1.3=py312h6c40b1e_0
mkl=2023.1.0=h8e150cf_43560
mkl-service=2.4.0=py312h6c40b1e_1
mkl_fft=1.3.8=py312h6c40b1e_0
mkl_random=1.2.4=py312ha357a0b_0
mpmath=1.3.0=py312hecd8cb5_0
ncurses=6.4=hcec6c5f_0
networkx=3.3=py312hecd8cb5_0
numpy=1.26.4=py312hac873b0_0
numpy-base=1.26.4=py312h6f81483_0
openjpeg=2.5.2=hbf2204d_0
openssl=3.0.15=h46256e1_0
packaging=24.1=py312hecd8cb5_0
pillow=10.4.0=py312h46256e1_0
pip=24.2=py312hecd8cb5_0
psutil=5.9.0=py312h6c40b1e_0
pysocks=1.7.1=py312hecd8cb5_0
python=3.12.7=hcd54a6c_0
pytorch=2.2.2=py3.12_0
pyyaml=6.0.2=py312h46256e1_0
readline=8.2=hca72f7f_0
regex=2024.9.11=py312h46256e1_0
requests=2.32.3=py312hecd8cb5_1
safetensors=0.4.5=py312hbc49121_0
setuptools=75.1.0=py312hecd8cb5_0
sqlite=3.45.3=h6c40b1e_0
sympy=1.13.2=py312hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
tk=8.6.14=h4d00af3_0
torchaudio=2.2.2=py312_cpu
torchvision=0.17.2=py312_cpu
tqdm=4.66.5=py312h8e4b320_0
typing-extensions=4.11.0=py312hecd8cb5_0
typing_extensions=4.11.0=py312hecd8cb5_0
tzdata=2024b=h04d1e81_0
urllib3=2.2.3=py312hecd8cb5_0
wheel=0.44.0=py312hecd8cb5_0
xz=5.4.6=h6c40b1e_1
yaml=0.2.5=haf1e3a3_0
zipp=3.20.2=py312hecd8cb5_0
zlib=1.2.13=h4b97444_1
zstd=1.5.6=h138b38a_0
